Move classes to tcp package
