+++ spring-cloud-starter-stream-processor-python-local/src/main/java/org/springframework/cloud/stream/app/python/local/tcp/TcpProcessor.java
@@ -1,6 +1,21 @@
-package org.springframework.cloud.stream.app.python.local.processor;
+/*
+ * Copyright 2017 the original author or authors.
+ *
+ *   Licensed under the Apache License, Version 2.0 (the "License");
+ *   you may not use this file except in compliance with the License.
+ *   You may obtain a copy of the License at
+ *
+ *        http://www.apache.org/licenses/LICENSE-2.0
+ *
+ *   Unless required by applicable law or agreed to in writing, software
+ *   distributed under the License is distributed on an "AS IS" BASIS,
+ *   WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *   See the License for the specific language governing permissions and
+ *   limitations under the License.
+ */
 
-import org.springframework.cloud.stream.app.python.jython.JythonScriptExecutor;
+package org.springframework.cloud.stream.app.python.local.tcp;
+
 import org.springframework.http.MediaType;
 import org.springframework.integration.ip.tcp.TcpOutboundGateway;
 import org.springframework.integration.support.MutableMessageBuilder;
